pyproject.toml
[project]
name = "advanced-unit-converter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.3.0",
    "pyht>=0.1.12",
    "pyttsx3>=2.98",
    "streamlit>=1.42.2",
]
